demos/streamll-rag-local/pyproject.toml
[project]
name = "streamll-rag-demo"
version = "0.1.0"
description = "Real-world RAG demo with StreamLL monitoring"
dependencies = [
    "streamll", # Our main package
    "dspy-ai",
    "llama-index",
    "llama-index-vector-stores-postgres",
    "llama-index-embeddings-google-genai",
    "psycopg2-binary",
    "redis",
    "python-dotenv",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
